moved App.tsx file back into components folder
